update: support both image and video collections

+++ app/src/main/java/com/fpf/smartscan/ui/components/collections/CollectionsActionBar.kt
@@ -36,29 +36,32 @@ fun CollectionsActionBar(
             .fillMaxWidth()
             .background(MaterialTheme.colorScheme.surfaceContainerLowest)
             .padding(horizontal = 24.dp),
-        horizontalArrangement = Arrangement.SpaceBetween,
+        horizontalArrangement = Arrangement.SpaceEvenly,
         verticalAlignment = Alignment.CenterVertically
     ) {
-        Button (
-            enabled = mergeEnabled,
-            colors = ButtonDefaults.buttonColors(
-                containerColor = Color.Transparent,
-                disabledContainerColor = Color.Transparent,
-                contentColor = MaterialTheme.colorScheme.onSurface
-            ),
-            onClick = { onMerge() }
-        ) {
-            Column(
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally) {
-                Icon(
-                    Icons.Filled.Merge,
-                    contentDescription = "Merge",
-                )
-                Text("Merge", style = MaterialTheme.typography.labelMedium)
+        if (mergeEnabled) {
+            Button(
+                enabled = mergeEnabled,
+                colors = ButtonDefaults.buttonColors(
+                    containerColor = Color.Transparent,
+                    disabledContainerColor = Color.Transparent,
+                    contentColor = MaterialTheme.colorScheme.onSurface
+                ),
+                onClick = { onMerge() }
+            ) {
+                Column(
+                    verticalArrangement = Arrangement.Center,
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Icon(
+                        Icons.Filled.Merge,
+                        contentDescription = "Merge",
+                    )
+                    Text("Merge", style = MaterialTheme.typography.labelMedium)
+                }
             }
         }
-        Button (
+        Button(
             enabled = renameEnabled,
             colors = ButtonDefaults.buttonColors(
                 containerColor = Color.Transparent,
@@ -69,7 +72,8 @@ fun CollectionsActionBar(
         ) {
             Column(
                 verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally){
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
                 Icon(
                     Icons.Filled.DriveFileRenameOutline,
                     contentDescription = "Rename",
@@ -77,23 +81,26 @@ fun CollectionsActionBar(
                 Text("Rename", style = MaterialTheme.typography.labelMedium)
             }
         }
-        Button (
-            enabled = deleteEnabled,
-            colors = ButtonDefaults.buttonColors(
-                containerColor = Color.Transparent,
-                disabledContainerColor = Color.Transparent,
-                contentColor = MaterialTheme.colorScheme.onSurface
-            ),
-            onClick = { onDelete() }
-        ) {
-            Column(
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally){
-                Icon(
-                    Icons.Filled.Delete,
-                    contentDescription = "Delete collection",
-                )
-                Text("Delete", style = MaterialTheme.typography.labelMedium)
+        if (deleteEnabled) {
+            Button(
+                enabled = deleteEnabled,
+                colors = ButtonDefaults.buttonColors(
+                    containerColor = Color.Transparent,
+                    disabledContainerColor = Color.Transparent,
+                    contentColor = MaterialTheme.colorScheme.onSurface
+                ),
+                onClick = { onDelete() }
+            ) {
+                Column(
+                    verticalArrangement = Arrangement.Center,
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Icon(
+                        Icons.Filled.Delete,
+                        contentDescription = "Delete collection",
+                    )
+                    Text("Delete", style = MaterialTheme.typography.labelMedium)
+                }
             }
         }
     }
